% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges.R
\name{add.edges.networkLite}
\alias{add.edges.networkLite}
\alias{[<-.networkLite}
\title{Methods to Add or Modify Edges in a \code{networkLite}}
\usage{
\method{add.edges}{networkLite}(x, tail, head, names.eval = NULL, vals.eval = NULL, ...)

\method{[}{networkLite}(x, i, j, names.eval = NULL, add.edges = FALSE) <- value
}
\arguments{
\item{x}{A \code{networkLite}.}

\item{tail}{Vector of tails of edges to add to the networkLite.}

\item{head}{Vector of heads of edges to add to the networkLite.}

\item{names.eval}{name(s) of edge attributes}

\item{vals.eval}{value(s) of edge attributes}

\item{...}{additional arguments}

\item{i, j}{Nodal indices (must be missing for networkLite method).}

\item{add.edges}{should edges being assigned to be added if not already
present?}

\item{value}{edge values to assign}
}
\description{
Methods to Add or Modify Edges in a \code{networkLite}
}
