% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vertices.R
\name{add.vertices.networkLite}
\alias{add.vertices.networkLite}
\title{Add Vertices to a \code{networkLite}}
\usage{
\method{add.vertices}{networkLite}(x, nv, vattr = NULL, last.mode = TRUE, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{nv}{Number of vertices to add to the \code{networkLite}.}

\item{vattr}{A list (of length \code{nv}) of named lists of vertex
attribute values for added vertices, or \code{NULL} to indicate vertex
attribute values are not being passed.}

\item{last.mode}{logical; if \code{x} is bipartite, should the new vertices
be added to the second mode?}

\item{...}{additional arguments}
}
\value{
A \code{networkLite} object with vertices added.
}
\description{
Add Vertices to a \code{networkLite}
}
