% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkLite.R
\name{as.networkLite}
\alias{as.networkLite}
\alias{as.networkLite.network}
\alias{as.networkLite.networkLite}
\title{as.networkLite}
\usage{
as.networkLite(x, ...)

\method{as.networkLite}{network}(x, ...)

\method{as.networkLite}{networkLite}(x, ...)
}
\arguments{
\item{x}{a \code{network} or \code{networkLite} object}

\item{...}{additional arguments}
}
\value{
a corresponding \code{networkLite} object
}
\description{
Convert to \code{networkLite} representation.
}
\details{
Currently the network attributes \code{hyper}, \code{multiple}, and
         \code{loops} must be \code{FALSE} for \code{networkLite}s;
         attempting to convert to \code{networkLite} when this is not the
         case will result in an error.
}
\seealso{
\code{\link{to_network_networkLite}}
}
