% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_edges.R
\name{delete.edges.networkLite}
\alias{delete.edges.networkLite}
\title{Delete edges from a networkLite.}
\usage{
\method{delete.edges}{networkLite}(x, eid, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{eid}{Edge ids (between \code{1} and
\code{network.edgecount(x, na.omit = FALSE)}) to delete in
\code{x}. Note that the edge id of an edge in \code{x} is simply
its row index in \code{x$el}.}
}
\description{
Delete edges from a networkLite.
}
