% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_vertices.R
\name{delete.vertices.networkLite}
\alias{delete.vertices.networkLite}
\title{Delete vertices from a networkLite.}
\usage{
\method{delete.vertices}{networkLite}(x, vid, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{vid}{Vertex ids (between \code{1} and \code{network.size(x)}) to delete
from \code{x}. Note that edges involving deleted vertices will
also be deleted.}

\item{...}{additional arguments.}
}
\description{
Delete vertices from a networkLite.
}
