% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_vertices.R
\name{get.inducedSubgraph.networkLite}
\alias{get.inducedSubgraph.networkLite}
\title{Return an induced subgraph}
\usage{
\method{get.inducedSubgraph}{networkLite}(x, v, alters = NULL, ...)
}
\arguments{
\item{x, v, alters, ...}{see \code{\link[network:get.inducedSubgraph]{network::get.inducedSubgraph()}}}
}
\description{
Return an induced subgraph
}
