% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.na.networkLite}
\alias{is.na.networkLite}
\title{Extract \code{networkLite} with Missing Edges Only}
\usage{
\method{is.na}{networkLite}(x)
}
\arguments{
\item{x}{A \code{networkLite}.}
}
\value{
A \code{networkLite} with the same network size, directedness, and
bipartiteness as \code{x}, whose edges are precisely those edges in
\code{x} that are missing in \code{x}. Edges in the returned
\code{networkLite} are marked as not missing.
}
\description{
Extract \code{networkLite} with Missing Edges Only
}
