% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.networkLite}
\alias{is.networkLite}
\title{Test whether the object is a \code{\link{networkLite}}}
\usage{
is.networkLite(x)
}
\arguments{
\item{x}{an object}
}
\description{
Test whether the object is a \code{\link{networkLite}}
}
