% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_network_networkLite.R
\name{$.networkLite}
\alias{$.networkLite}
\alias{[.networkLite}
\title{Last-resort utilities}
\usage{
\method{$}{networkLite}(x, name)

\method{[}{networkLite}(x, ...)
}
\arguments{
\item{x}{a \code{networkLite} object}

\item{name}{passed to \code{[[}, but some special cases (\code{"oel"} nd \code{"iel"} are handled)}
}
\description{
Since the purpose of \code{\link{networkLite}} is to be a drop-in replacement
for \code{\link[network:network]{network}}, these utilities emulate the internal structure of
\code{network}. They are likely to perform poorly and should be avoided.
}
\note{
Please do not rely on these unless you absolutely have to;
since they slow down access to \code{networkLite} objects, they may be
removed if they ever stop being necessary.
}
