% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixingmatrix.R
\name{mixingmatrix.networkLite}
\alias{mixingmatrix.networkLite}
\title{Extract Mixing Matrix from \code{networkLite}}
\usage{
\method{mixingmatrix}{networkLite}(object, attr, ...)
}
\arguments{
\item{object}{A \code{networkLite} object.}

\item{attr}{The name of a vertex attribute in \code{object}.}

\item{...}{additional arguments}
}
\value{
The mixing matrix for \code{object} and \code{attr}.
}
\description{
Extract Mixing Matrix from \code{networkLite}
}
