% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{+.networkLite}
\alias{+.networkLite}
\alias{-.networkLite}
\title{Add and Subtract \code{networkLite}s}
\usage{
\method{+}{networkLite}(e1, e2)

\method{-}{networkLite}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{networkLite} objects}
}
\description{
Add and Subtract \code{networkLite}s
}
\details{
\code{e1 + e2} produces a \code{networkLite} whose edges are those
in either \code{e1} or \code{e2}, and \code{e1 - e2} produces a
\code{networkLite} whose edges are those in \code{e1} and not in
\code{e2}.
}
