% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vertices.R
\name{permute.vertexIDs.networkLite}
\alias{permute.vertexIDs.networkLite}
\title{Permute vertices}
\usage{
\method{permute.vertexIDs}{networkLite}(x, vids, ...)
}
\arguments{
\item{x, vids}{see \code{\link[network:permute.vertexIDs]{network::permute.vertexIDs()}}}
}
\description{
Permute vertices
}
