% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.networkLite}
\alias{print.networkLite}
\title{Print Basic Summary of a \code{networkLite}}
\usage{
\method{print}{networkLite}(x, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{...}{additional arguments}
}
\value{
The \code{networkLite} is returned invisibly.
}
\description{
Print Basic Summary of a \code{networkLite}
}
\details{
This method prints a basic summary of a \code{networkLite} object,
including network size, edge count, and attribute names.
}
