% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkLite.R
\name{to_network_networkLite}
\alias{to_network_networkLite}
\title{Convert networkLite to network}
\usage{
to_network_networkLite(x, ...)
}
\arguments{
\item{x}{a \code{networkLite} object}

\item{...}{additional arguments}
}
\value{
a corresponding \code{network} object
}
\description{
Convert networkLite to network
}
