% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.na.networkLite}
\alias{is.na.networkLite}
\title{Extract \code{networkLite} with Missing Edges Only}
\usage{
\method{is.na}{networkLite}(x)
}
\arguments{
\item{x}{A \code{networkLite}.}
}
\description{
Extract \code{networkLite} with Missing Edges Only
}
\details{
The \code{is.na} method creates a \code{networkLite} whose edges
         are precisely those edges in \code{x} that are missing.  The edges
         in the return value are marked as not missing.
}
