% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_network_networkLite.R
\name{to_network_networkLite}
\alias{to_network_networkLite}
\alias{as.network.networkLite}
\title{Convert a \code{networkLite} object to a \code{network} object}
\usage{
to_network_networkLite(x, ...)

\method{as.network}{networkLite}(x, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{...}{additional arguments}
}
\value{
A corresponding \code{network} object.
}
\description{
Convert a \code{networkLite} object to a \code{network} object
}
\details{
The \code{to_network_networkLite} function takes a
\code{networkLite} and returns a corresponding \code{network}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     The \code{as.network.networkLite} method returns the
     \code{networkLite} unchanged, for compatibility with \code{ergm}.
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link{as.networkLite}}
}
